+++ src/java.base/share/classes/java/lang/foreign/Linker.java
@@ -168,7 +168,8 @@ public sealed interface Linker
      */
     default MethodHandle downcallHandle(MemorySegment symbol, FunctionDescriptor function) {
         SharedUtils.checkSymbol(symbol);
-        return downcallHandle(function).bindTo(symbol);
+        return downcallHandle(function)
+                .bindTo(symbol);
     }
 
     /**
