+++ src/java.base/share/classes/java/lang/foreign/MemoryLayout.java
@@ -33,7 +33,6 @@ import java.util.EnumSet;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
-import java.util.function.Function;
 import java.util.function.Supplier;
 import java.util.stream.Stream;
 import jdk.internal.foreign.LayoutPath;
@@ -259,7 +258,7 @@ public sealed interface MemoryLayout
      * in {@code elements} is {@code null}.
      */
     default long byteOffset(PathElement... elements) {
-        return Utils.bitsToBytesOrThrow(bitOffset(elements), Utils.bitsToBytesThrowOffset);
+        return Utils.bitsToBytesOrThrow(bitOffset(elements), Utils.BITS_TO_BYTES_THROW_OFFSET);
     }
 
     /**
@@ -275,23 +274,10 @@ public sealed interface MemoryLayout
      * in {@code elements} is {@code null}.
      */
     default long bitOffset(PathElement... elements) {
-        return computePathOp(LayoutPath.rootPath(this), LayoutPath::offset,
-                EnumSet.of(PathKind.SEQUENCE_ELEMENT, PathKind.SEQUENCE_RANGE), elements);
-    }
-
-    private static <Z> Z computePathOp(LayoutPath path,
-                                       Function<LayoutPath, Z> finalizer,
-                                       Set<PathKind> badKinds,
-                                       PathElement... elements) {
-        Objects.requireNonNull(elements);
-        for (PathElement e : elements) {
-            LayoutPath.PathElementImpl pathElem = (LayoutPath.PathElementImpl)Objects.requireNonNull(e);
-            if (badKinds.contains(pathElem.kind())) {
-                throw new IllegalArgumentException(String.format("Invalid %s selection in layout path", pathElem.kind().description()));
-            }
-            path = pathElem.apply(path);
-        }
-        return finalizer.apply(path);
+        return Utils.computePathOp(LayoutPath.rootPath(this),
+                LayoutPath::offset,
+                EnumSet.of(PathKind.SEQUENCE_ELEMENT, PathKind.SEQUENCE_RANGE),
+                elements);
     }
 
     /**
@@ -327,7 +313,7 @@ public sealed interface MemoryLayout
      */
     default MethodHandle byteOffsetHandle(PathElement... elements) {
         MethodHandle mh = bitOffsetHandle(elements);
-        mh = MethodHandles.filterReturnValue(mh, Utils.MH_bitsToBytesOrThrowForOffset);
+        mh = MethodHandles.filterReturnValue(mh, Utils.MH_BITS_TO_BYTES_OR_THROW_FOR_OFFSET);
         return mh;
     }
 
@@ -359,8 +345,10 @@ public sealed interface MemoryLayout
      * multiple sequence element indices (see {@link PathElement#sequenceElement(long, long)}).
      */
     default MethodHandle bitOffsetHandle(PathElement... elements) {
-        return computePathOp(LayoutPath.rootPath(this), LayoutPath::offsetHandle,
-                EnumSet.of(PathKind.SEQUENCE_RANGE), elements);
+        return Utils.computePathOp(LayoutPath.rootPath(this),
+                LayoutPath::offsetHandle,
+                EnumSet.of(PathKind.SEQUENCE_RANGE),
+                elements);
     }
 
     /**
@@ -399,8 +387,10 @@ public sealed interface MemoryLayout
      * @see MethodHandles#memorySegmentViewVarHandle(ValueLayout)
      */
     default VarHandle varHandle(PathElement... elements) {
-        return computePathOp(LayoutPath.rootPath(this), LayoutPath::dereferenceHandle,
-                Set.of(), elements);
+        return Utils.computePathOp(LayoutPath.rootPath(this),
+                LayoutPath::dereferenceHandle,
+                Set.of(),
+                elements);
     }
 
     /**
@@ -442,8 +432,10 @@ public sealed interface MemoryLayout
      * @throws UnsupportedOperationException if the size of the selected layout in bits is not a multiple of 8.
      */
     default MethodHandle sliceHandle(PathElement... elements) {
-        return computePathOp(LayoutPath.rootPath(this), LayoutPath::sliceHandle,
-                Set.of(), elements);
+        return Utils.computePathOp(LayoutPath.rootPath(this),
+                LayoutPath::sliceHandle,
+                Set.of(),
+                elements);
     }
 
     /**
@@ -456,14 +448,11 @@ public sealed interface MemoryLayout
      * (see {@link PathElement#sequenceElement(long)} and {@link PathElement#sequenceElement(long, long)}).
      */
     default MemoryLayout select(PathElement... elements) {
-        return computePathOp(LayoutPath.rootPath(this), LayoutPath::layout,
-                EnumSet.of(PathKind.SEQUENCE_ELEMENT_INDEX, PathKind.SEQUENCE_RANGE), elements);
+        return Utils.computePathOp(LayoutPath.rootPath(this),
+                LayoutPath::layout,
+                EnumSet.of(PathKind.SEQUENCE_ELEMENT_INDEX, PathKind.SEQUENCE_RANGE),
+                elements);
     }
-
-    /**
-     * {@return the hash code value for this layout}
-     */
-    int hashCode();
 
     /**
      * Compares the specified object with this layout for equality. Returns {@code true} if and only if the specified
@@ -485,12 +474,6 @@ public sealed interface MemoryLayout
     boolean equals(Object other);
 
     /**
-     * {@return the string representation of this layout}
-     */
-    @Override
-    String toString();
-
-    /**
      * Creates a padding layout with the given size.
      *
      * @param size the padding size in bits.
@@ -500,51 +483,6 @@ public sealed interface MemoryLayout
     static PaddingLayout paddingLayout(long size) {
         MemoryLayoutUtil.checkSize(size);
         return PaddingLayoutImpl.of(size);
-    }
-
-    /**
-     * Creates a value layout of given Java carrier and byte order. The type of resulting value layout is determined
-     * by the carrier provided:
-     * <ul>
-     *     <li>{@link ValueLayout.OfBoolean}, for {@code boolean.class}</li>
-     *     <li>{@link ValueLayout.OfByte}, for {@code byte.class}</li>
-     *     <li>{@link ValueLayout.OfShort}, for {@code short.class}</li>
-     *     <li>{@link ValueLayout.OfChar}, for {@code char.class}</li>
-     *     <li>{@link ValueLayout.OfInt}, for {@code int.class}</li>
-     *     <li>{@link ValueLayout.OfFloat}, for {@code float.class}</li>
-     *     <li>{@link ValueLayout.OfLong}, for {@code long.class}</li>
-     *     <li>{@link ValueLayout.OfDouble}, for {@code double.class}</li>
-     *     <li>{@link ValueLayout.OfAddress}, for {@code MemorySegment.class}</li>
-     * </ul>
-     * @param carrier the value layout carrier.
-     * @param order the value layout's byte order.
-     * @return a value layout with the given Java carrier and byte-order.
-     * @throws IllegalArgumentException if the carrier type is not supported.
-     */
-    static ValueLayout valueLayout(Class<?> carrier, ByteOrder order) {
-        Objects.requireNonNull(carrier);
-        Objects.requireNonNull(order);
-        if (carrier == boolean.class) {
-            return ValueLayouts.OfBooleanImpl.of(order);
-        } else if (carrier == char.class) {
-            return ValueLayouts.OfCharImpl.of(order);
-        } else if (carrier == byte.class) {
-            return ValueLayouts.OfByteImpl.of(order);
-        } else if (carrier == short.class) {
-            return ValueLayouts.OfShortImpl.of(order);
-        } else if (carrier == int.class) {
-            return ValueLayouts.OfIntImpl.of(order);
-        } else if (carrier == float.class) {
-            return ValueLayouts.OfFloatImpl.of(order);
-        } else if (carrier == long.class) {
-            return ValueLayouts.OfLongImpl.of(order);
-        } else if (carrier == double.class) {
-            return ValueLayouts.OfDoubleImpl.of(order);
-        } else if (carrier == MemorySegment.class) {
-            return ValueLayouts.OfAddressImpl.of(order);
-        } else {
-            throw new IllegalArgumentException("Unsupported carrier: " + carrier.getName());
-        }
     }
 
     /**
